refactor: move markdownToText function to markdownToJsx.js module
